requirements.txt
requests
websocket-client
pandas
numpy
plyer
protobuf
google
Flask
Flask-SocketIO
python-dotenv
eventlet